pyproject.toml
[project]
name = "sds"
version = "0.1.0"
description = "The SuperDataScience AI Bootcamp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.53.0",
    "chromadb>=1.0.12",
    "docker>=7.1.0",
    "gradio>=5.33.1",
    "ipywidgets>=8.1.7",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-openai>=0.3.22",
    "langchain-text-splitters>=0.3.8",
    "litellm>=1.72.6",
    "mcp[cli]>=1.9.3",
    "nbformat>=5.10.4",
    "openai>=1.86.0",
    "openai-agents[litellm,viz,voice]>=0.0.17",
    "playwright>=1.52.0",
    "plotly>=6.1.2",
    "psutil>=7.0.0",
    "pypdf>=5.6.0",
    "pysonar>=1.1.0.2035",
    "python-dotenv>=1.1.0",
    "ratelimit>=2.2.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
